Первою ревью

+++ controllers/cards.js
@@ -1,73 +1,66 @@
+const { ValidationError } = require('mongoose').Error;
+const BadRequestError = require('../errors/BadRequestError');
+const ForbiddenError = require('../errors/ForbiddenError');
+const NotFoundError = require('../errors/NotFoundError');
 const Card = require('../models/card');
 const {
-  INTERNAL_SERVER_ERROR_STATUS,
-  BAD_REQUEST_ERROR_STATUS,
-  NOT_FOUND_ERROR_STATUS,
-  CREATE_STATUS,
-} = require('../utils/constantStatusCode');
+CREATE_STATUS,
+} = require('../utils/constants');
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
+    .sort({ createdAt: -1 })
+    .populate(['owner', 'likes'])
     .then((cards) => res.send({ data: cards }))
-    .catch(() => {
-      res.status(INTERNAL_SERVER_ERROR_STATUS).send({ message: 'Произошла ошибка' });
-    });
+    .catch(next);
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   Card.create({ name, link, owner: req.user._id })
-    .then((cards) => res.status(CREATE_STATUS).send({ data: cards }))
+    .then((card) => card.populate('owner'))
+    .then((card) => res.status(CREATE_STATUS).send({ data: card }))
     .catch((err) => {
-      if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST_ERROR_STATUS).send({ message: 'Переданы некорректные данные' });
+      if (err instanceof ValidationError) {
+        next(new BadRequestError('Переданы некорректные данные'));
       } else {
-        res.status(INTERNAL_SERVER_ERROR_STATUS).send({ message: 'Ошибка сервера' });
+        next(err);
       }
     });
 };
 
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((cards) => {
-      if (!cards) {
-        res.status(NOT_FOUND_ERROR_STATUS).send({ message: 'Карточка не найдена' });
-        return;
+const deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Карточка не найдена');
+      } else if (req.user._id === card.owner.toString()) {
+        return Card.findByIdAndRemove(req.params.cardId)
+          .then(() => res.send({ message: 'Карточка удалена' }));
+      } else {
+        return next(new ForbiddenError('Нельзя удалять не ваши карточки'));
       }
-      res.send({ data: cards });
     })
-    .catch((err) => {
-      if (err.name === 'CastError' && err.kind === 'ObjectId') {
-        res.status(BAD_REQUEST_ERROR_STATUS).send({ message: 'Неверный формат ID карточки' });
-      } else {
-        res.status(INTERNAL_SERVER_ERROR_STATUS).send({ message: 'Произошла ошибка' });
-      }
-    });
+    .catch(next);
 };
 
-const likeCard = (req, res) => {
+const likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
+    .populate(['owner', 'likes'])
     .then((card) => {
       if (!card) {
-        res.status(NOT_FOUND_ERROR_STATUS).send({ message: 'Карточка не найдена' });
-        return;
+        throw new NotFoundError('Карточка не найдена');
       }
       res.send(card);
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(BAD_REQUEST_ERROR_STATUS).send({ message: 'Неверный формат ID карточки' });
-      } else {
-        res.status(INTERNAL_SERVER_ERROR_STATUS).send({ message: 'Ошибка сервера' });
-      }
-    });
+    .catch(next);
 };
 
-const dislikeCard = (req, res) => {
+const dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -75,18 +68,11 @@ const dislikeCard = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        res.status(NOT_FOUND_ERROR_STATUS).send({ message: 'Карточка не найдена' });
-        return;
+        throw new NotFoundError('Карточка не найдена');
       }
       res.send(card);
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(BAD_REQUEST_ERROR_STATUS).send({ message: 'Неверный формат ID карточки' });
-      } else {
-        res.status(INTERNAL_SERVER_ERROR_STATUS).send({ message: 'Ошибка сервера' });
-      }
-    });
+    .catch(next);
 };
 
 module.exports = {
